feat(cli): refactor logic into classes
